chore(meta): move FetchAddU64 utils to separate file
